orth: Make pulling the docker image configurable
Extract variables for hard-coded configuration and move them to the
configuration file. As authentication is not required by all Docker
registries, perform the authentication only in case the registry server
and credentials are providided.

Change the registry server to an empty string because authentication is
not required for pulling the specified Docker image.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>