Drop the import/export example files
These files do not fit into the 'ort-config' repository as they are not
part of the ORT configuration. The file `license-finding-curations.yml`
and `path-excludes.yml` have been generated by the `*-export-*`
commands. So, there is no need to examplify these files because the user
can generate them. Also the entries of the mentioned files are not
useful as-is. Finally, `vcs-url-mapping` is a trivial configuration of
the `import` and `export` commands. This configuration could be better
illustrated in a documentation file like README.md.

So, drop all the files as they are useless.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>