orthw: Fix the broken `check_evaluation_md5_sum()`
The MD5 checksum of the package curations was properly handled in
`update_evaluation_md5_sum()` but missing in
`check_evaluation_md5_sum()`. So, when recomputing the report after
making a file change in the package curations directory the evaluator
was not re-run. Add the missing check, analog to the one for package
configurations, to fix that problem.

Fixes #19.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>