orthw: Fix an issue with the init command
The init command intends to work for `*.json`, `*.json.xz`, `*.yml` and
`*.yml.xz` files. It works if the filename part matched by `*` does not
contain any dot, but otherwise it fails. Re-write the logic to properly
use the substring after the last dot for determining the extension, to
avoid messing up the value of `temp_scan_result_file`. This makes the
initialization work with any filenames matching the above mentioned
patterns.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>