feat(commands): Add delete-scan-results
Added first iteraction of delete-scan-results command.
Bandit checker need to be silenced due possibility of SQL injection
warnings, which is prevented with own psycopg2 literal prevention.

From original orthw there's no mention where the database configurations
are restored, so for now it assumes will be available on main config. A
possible solution need to be devised.

Signed-off-by: Helio Chissini de Castro <heliocastro@gmail.com>