feat(orthw): Migrate the obsolete `(delete-)scan-results` commands
The previous commands, `delete-scan-results` and `scan-results` have
been targeting the following use case:

1. Scanning Git-Repo projects with non-backed manifests use a "wrong"
   cache key. The manual deletion of scan results is a work around to
   force a re-scan.
2. If ORT is configured to scan source code origins `{ VCS, ARTIFACT }`,
   and it has a scan result for `ARTIFACT` in the scan storage,
   deleting the result after setting proper VCS info will make ORT
   use the preferred source code origin instead of the fallback one.

Nowadays, the scan-results` table has become obsolete and the relation
between a package and a scan results is represented by the tables
`package_provenances`, `nested_provenances` and
`provenances_scan_results`. In order to re-enable the above use cases it
suffices to just delete the entry in the `package_provenances` table.
So, adjust the commands to operate on package provenances instead of
scan results.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>